Removed chunk adding code (Need Fixing), cleaned up some code in other areas

+++ Vulkan API/src/game/TerrainGeneration/TerrainGenerator.h
@@ -20,7 +20,7 @@ namespace Vulkan
 		float interpolate(float a0, float a1, float w);
 		
 		std::vector<std::pair<glm::vec2, glm::vec2>> RandomChunkBorderGradients;
-		float seed;
+		int seed;
 		int chunk_scale_;
 	};
 }
